+++ autoload/omnipytent/tasks.py
@@ -4,7 +4,7 @@ import vim
 
 from .context import InvocationContext
 from .hacks import function_locals
-from .util import input_list, other_windows
+from .util import input_list, other_windows, flatten_iterator
 
 
 class Task(object):
@@ -19,12 +19,24 @@ class Task(object):
         self.dependencies = []
         self.completers = []
 
+    def _run_func_as_generator(self, *args, **kwargs):
+        result = self.func(*args, **kwargs)
+        if inspect.isgenerator(result):
+            result = flatten_iterator(result)
+            try:
+                yielded = next(result)
+                while True:
+                    yield yielded
+                    yielded = result.send(yielded._returned_value)
+            except StopIteration:
+                pass
+
     def invoke(self, ctx, *args):
         ctx = ctx.for_task(self)
-        if ctx.is_main:
-            self.func(ctx, *args)
-        else:
-            self.func(ctx)
+        if not ctx.is_main:
+            args = ()
+        for yielded in self._run_func_as_generator(ctx, *args):
+            yield yielded
 
     def __repr__(self):
         return '<Task: %s>' % self.name
@@ -63,13 +75,14 @@ class OptionsTask(Task):
             return []
 
     def invoke(self, ctx, *args):
+        from .async_execution import CHOOSE
+        if False:
+            yield  # force this into a genrator
         ctx = ctx.for_task(self)
         if len(self.__func_args_set) == 0:
             options = function_locals(self.func)
         else:
             options = function_locals(self.func, ctx)
-
-        # options = {k: v for k, v in options.items() if self.__varname_filter(k)}
 
         if 0 == len(args) or not ctx.is_main:
             if ctx.is_main:
@@ -84,7 +97,10 @@ class OptionsTask(Task):
                 elif 1 == len(options):
                     ctx.pass_data(next(iter(options.values())))
                     return
-                chosen_item = input_list('Choose %s' % self, options_keys)
+                # chosen_item = yield CHOOSE(options_keys, prompt='Choose %s' % self)
+                async_cmd = CHOOSE(options_keys)
+                yield async_cmd
+                chosen_item = async_cmd._returned_value
                 if chosen_item:
                     ctx.cache.chosen_item = chosen_item
 
@@ -131,7 +147,8 @@ class WindowTask(Task):
             pass
 
         with other_windows():
-            super(WindowTask, self).invoke(ctx, *args)
+            for yielded in super(WindowTask, self).invoke(ctx, *args):
+                yield yielded
             window = vim.current.window
         task_ctx.cache.window = window
         if task_ctx.has_passed_data:
@@ -155,4 +172,5 @@ def invoke_with_dependencies(tasks_file, task, args):
         for task in run_order:
             if task not in already_invoked:
                 already_invoked.add(task)
-                task.invoke(ctx, *args)
+                for yielded in task.invoke(ctx, *args):
+                    yield yielded