+++ autoload/omnipytent/util.py
@@ -3,9 +3,26 @@ import vim
 from contextlib import contextmanager
 import json
 import re
+import os.path
+import inspect
+
+
+class RawVim(str):
+    @classmethod
+    def fmt(cls, fmt, *args):
+        return cls(fmt % args)
 
 
 def vim_repr(value):
+    if value is None:
+        return '0'
+
+    if isinstance(value, RawVim):
+        return value
+
+    if hasattr(value, '_to_raw_vim_'):
+        return value._to_raw_vim_()
+
     if isinstance(value, bool):
         return str(int(value))  # 0 or 1
 
@@ -31,7 +48,7 @@ def vim_repr(value):
     if isinstance(value, (str, bytes)):
         return json.dumps(value)
 
-    if isinstance(value, (list, set)):
+    if isinstance(value, (list, set, tuple)):
         return '[%s]' % ', '.join(map(vim_repr, value))
 
     if isinstance(value, dict):
@@ -46,17 +63,30 @@ __number_type = vim.eval('type(0)')
 __float_type = vim.eval('type(0.0)')
 
 
+def __apply_type_map(value, type_map):
+    if type_map == __number_type:
+        return int(value)
+    elif type_map == __float_type:
+        return float(value)
+    elif isinstance(type_map, list):
+        assert isinstance(value, list)
+        assert len(value) == len(type_map)
+        return [__apply_type_map(v, t) for v, t in zip(value, type_map)]
+    elif isinstance(type_map, dict):
+        assert isinstance(value, dict)
+        assert value.keys() == type_map.keys()
+        return {k: __apply_type_map(v, type_map[k]) for k, v in value.items()}
+    else:
+        return value
+
+
 def vim_eval(expr):
     """Like vim.eval, but deals with numbers correctly"""
-    expr_type, expr_value = vim.eval('map([%s], "[type(v:val), v:val]")[0]' % expr)
-    if expr_type == __number_type:
-        expr_value = int(expr_value)
-    elif expr_type == __float_type:
-        expr_value = float(expr_value)
-    return expr_value
+    # expr_type, expr_value = vim.eval('map([%s], "[type(v:val), v:val]")[0]' % expr)
+    return __apply_type_map(*vim.eval('omnipytent#_typeMap(%s)' % (expr,)))
 
 
-def input_list(prompt, options):
+def input_list(prompt, options, fmt=str):
     take_from = 0
     while take_from < len(options):
         take_this_time = int(vim.eval('&lines')) - 2
@@ -81,7 +111,7 @@ def input_list(prompt, options):
                 index_text = str(index + 1)
                 yield '%s)%s %s' % (index_text,
                                     ' ' * (number_length - len(index_text)),
-                                    option)
+                                    fmt(option))
         list_for_input_query = list(list_for_input_query_generator())
         chosen_option_number = int(vim.eval("inputlist(%s)" % vim_repr(list_for_input_query)))
 
@@ -156,3 +186,39 @@ def other_windows(window=None):
         vim.current.window = original_window
         if original_mode == 'n':
             vim.command(r'call feedkeys("\<C-\>\<C-n>")')
+
+
+def load_companion_vim_file(source_file):
+    """
+    Load a companion Vim file with the same name of a Python file.
+
+    For example, from ``foo.py`` call::
+
+        load_companion_vim_file(__name__)
+
+    To load ``foo.vim`` in the same directory.
+    """
+
+    path, ext = os.path.splitext(os.path.abspath(source_file))
+    path += '.vim'
+    vim.command('source ' + path)
+
+
+def flatten_iterator(it):
+    try:
+        item = next(it)
+        while True:
+            if inspect.isgenerator(item):
+                item = flatten_iterator(item)
+                try:
+                    subitem = next(item)
+                    while True:
+                        send = yield subitem
+                        subitem = item.send(send)
+                except StopIteration:
+                    pass
+            else:
+                send = yield item
+            item = it.send(send)
+    except StopIteration:
+        pass